# docs/swagger.yaml
definitions:
...
  entity.OrderInput:
    properties:
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      order_id:
        type: integer
      ordered_items:
        items:
          $ref: '#/definitions/entity.OrderedItemInput'
        type: array
      status:
        type: string
      updated_at:
        type: string
      warehouse_id:
        type: integer
    type: object
  entity.OrderedItemInput:
    properties:
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  entity.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /orderedItems:
    get:
      consumes:
      - application/json
      description: Retrieve all ordered items from the database
      produces:
      - application/json
      responses:
        "200":
          description: Ordered items getted
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetOrderedItems error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Retrieve all ordered items
      tags:
      - Ordered Item
  /orderedItems/{id}:
    get:
      consumes:
      - application/json
      description: Get OrderedItem details by Order ID
      parameters:
      - description: OrderedItem ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get OrderedItemsByOrderId
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid Order ID GetOrderedItemsByOrderId
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetOrderedItemsByOrderId error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get orderedItems
      tags:
      - OrderedItem
  /orders:
    post:
      consumes:
      - application/json
      description: Add an order to the database
      parameters:
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/entity.OrderInput'
      produces:
      - application/json
      responses:
        "201":
          description: Order created
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid order_id format, Unable to parse form data, Unable
            to get Order from form
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application AddOrder error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Add order
      tags:
      - Order
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an Order from the database by ID
      parameters:
      - description: order_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order deleted
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid order_id DeleteOrder
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application DeleteOrder error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Delete order
      tags:
      - Order
    get:
      consumes:
      - application/json
      description: Get Order details by order_id
      parameters:
      - description: order_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get Order
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid order_id GetOrder
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetOrder error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get order
      tags:
      - Order
    put:
      consumes:
      - application/json
      description: Update an order in the database by ID
      parameters:
      - description: order_id
        in: path
        name: id
        required: true
        type: integer
      - description: Order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/entity.OrderInput'
      produces:
      - application/json
      responses:
        "201":
          description: Order updated
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid order_id
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application UpdateOrder error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Update order
      tags:
      - Order
...
swagger: "2.0"